refactor: move env and logger to utils
