+++ packages/onegrep-api-client/src/index.ts
@@ -1,3 +1,2 @@
 export * from './api.js'
 export * from './client.js'
-export * from './env.js'
