+++ packages/onegrep-sdk/package.json
@@ -54,6 +54,7 @@
     "vitest": "^3.0.5"
   },
   "dependencies": {
+    "@repo/utils": "workspace:*",
     "@modelcontextprotocol/sdk": "^1.5.0",
     "json-schema-to-zod": "^2.6.0",
     "eventsource": "^3.0.5",
