+++ packages/utils/tsconfig.json
@@ -3,6 +3,6 @@
   "compilerOptions": {
     "rootDir": "src",
   },
-  "include": ["src/**/*", "tests/**/*"],
+  "include": ["src/**/*"],
   "exclude": ["node_modules"]
 }
